requirements.txt
numpy
pandas
scipy
scikit-learn
pyyaml
mne
torch
matplotlib
pytest
